fix ddl deploy script handling of new tables
Previously, when the DDL Deploy script encountered a table for
which the AWS Glue CLI returned no DDL update time, it would put
this table into the list of tables to update and then break the
loop over all the tables. This inadventently caused the script
to ignore changes to later tables in the loop.

This commit fixes the issue of the loop prematurely breaking and
also grabs the output from the AWS Glue CLI command for both
stdout and stderr. This avoids the AWS Glue CLI from reporting
its errors into the deployment output, which could possibly
confuse users reading the output. There will be expected errors
returned from the AWS Glue CLI when we query for tables that
have not yet been deployed.